🌱 Split kubectlProxy.additional.test.ts into 3 focused test files
Split 2465-line kubectlProxy.additional.test.ts into:
- kubectlProxy.additional.connection.test.ts (~467 lines): parseResourceQuantity, ensureConnected, execImmediate, processQueue
- kubectlProxy.additional.resources.test.ts (~913 lines): getServices/getPVCs/getPodIssues/getEvents/getDeployments/getClusterUsage/getClusterHealth/getNodes/getPodMetrics error branches
- kubectlProxy.additional.edge.test.ts (~1500 lines): getBulkClusterHealth, non-Error handling, missing items, edge cases

All 77 tests pass. Part of #10439.
