+++ web/src/lib/__tests__/kubectlProxy.additional.edge.test.ts
@@ -204,973 +204,8 @@ async function createProxy() {
 // functions identified by coverage analysis
 // ===========================================================================
 
-describe('KubectlProxy — additional coverage', () => {
-  // =========================================================================
-  // parseResourceQuantity: SI suffixes K, M, G, T and edge cases
-  // =========================================================================
 
-  describe('parseResourceQuantity — SI decimal suffixes (K, M, G, T)', () => {
-    /** Helper: parse a memory/storage value via getNodes */
-    async function parseViaNodeMemory(value: string): Promise<number> {
-      const proxy = await createProxy()
-      const nodesPromise = proxy.getNodes('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: {
-          output: JSON.stringify({
-            items: [{
-              metadata: { name: 'n1', labels: {} },
-              status: {
-                conditions: [{ type: 'Ready', status: 'True' }],
-                allocatable: { memory: value },
-              },
-            }],
-          }),
-          exitCode: 0,
-        },
-      })
-      const nodes = await nodesPromise
-      proxy.close()
-      return nodes[0].memoryBytes!
-    }
-
-    it('parses K (kilobytes, decimal)', async () => {
-      expect(await parseViaNodeMemory('10K')).toBe(10_000)
-    })
-
-    it('parses M (megabytes, decimal)', async () => {
-      expect(await parseViaNodeMemory('5M')).toBe(5_000_000)
-    })
-
-    it('parses G (gigabytes, decimal)', async () => {
-      expect(await parseViaNodeMemory('2G')).toBe(2_000_000_000)
-    })
-
-    it('parses T (terabytes, decimal)', async () => {
-      expect(await parseViaNodeMemory('1T')).toBe(1_000_000_000_000)
-    })
-
-    it('returns 0 for completely unparseable input', async () => {
-      expect(await parseViaNodeMemory('not-a-number')).toBe(0)
-    })
-
-    it('falls back to parseFloat for suffix-less numeric string', async () => {
-      expect(await parseViaNodeMemory('12345')).toBe(12345)
-    })
-  })
-
-  // =========================================================================
-  // ensureConnected: isConnecting wait-and-retry branch (lines 91-94)
-  // =========================================================================
-
-  describe('ensureConnected — isConnecting guard', () => {
-    it('waits and retries when another connection attempt is in progress', async () => {
-      const proxy = await createProxy()
-
-      // Start two exec calls nearly simultaneously — the second will hit the
-      // isConnecting guard because connectPromise is set to null during the
-      // brief window between setting isConnecting=true and assigning connectPromise.
-      // We simulate this by starting the first exec, then before open fires,
-      // immediately starting another exec.
-
-      const exec1 = proxy.exec(['get', 'pods'], { priority: true })
-      await vi.advanceTimersByTimeAsync(0)
-
-      // First WS is created; its connectPromise exists.
-      // Now start a second exec while the first is still connecting.
-      const exec2 = proxy.exec(['get', 'nodes'], { priority: true })
-      await vi.advanceTimersByTimeAsync(0)
-
-      // Open the WS — both should now be able to proceed
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      // There should now be 2 messages sent
-      expect(sentMessages.length).toBe(2)
-
-      // Respond to both
-      for (const rawMsg of sentMessages) {
-        const msg = JSON.parse(rawMsg)
-        activeWs!.simulateMessage({
-          id: msg.id,
-          type: 'result',
-          payload: { output: 'ok', exitCode: 0 },
-        })
-      }
-
-      await Promise.all([exec1, exec2])
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // ensureConnected: WebSocket constructor throws (lines 165-168)
-  // =========================================================================
-
-  describe('ensureConnected — WebSocket constructor throws', () => {
-    it('covers the catch block and falls through to "Not connected" guard', async () => {
-      const proxy = await createProxy()
-      // Replace WebSocket with a constructor that throws — the catch block
-      // in ensureConnected nulls out connectPromise, so the error is caught
-      // but ensureConnected returns null; execImmediate then hits the
-      // "Not connected to local agent" guard.
-      vi.stubGlobal('WebSocket', class {
-        constructor() {
-          throw new Error('WebSocket not supported')
-        }
-      })
-
-      await expect(proxy.exec(['get', 'pods'])).rejects.toThrow('Not connected to local agent')
-
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // finalize double-call guard (line 102)
-  // =========================================================================
-
-  describe('ensureConnected — finalize double-call guard', () => {
-    it('ignores second settlement when both timeout and error fire', async () => {
-      const proxy = await createProxy()
-
-      const execPromise = proxy.exec(['get', 'pods'])
-      const rejection = expect(execPromise).rejects.toThrow()
-      await vi.advanceTimersByTimeAsync(0)
-
-      // Fire error first — this settles the promise
-      activeWs!.simulateError()
-      await vi.advanceTimersByTimeAsync(0)
-
-      // Now let the connect timeout also fire — it should be a no-op
-      await vi.advanceTimersByTimeAsync(2500)
-
-      await rejection
-
-      proxy.close()
-    })
-
-    it('ignores open after timeout already fired', async () => {
-      const proxy = await createProxy()
-
-      const execPromise = proxy.exec(['get', 'pods'])
-      const rejection = expect(execPromise).rejects.toThrow('Connection timeout')
-      await vi.advanceTimersByTimeAsync(0)
-
-      // Let timeout fire first
-      await vi.advanceTimersByTimeAsync(2500)
-
-      // Now simulate open — should be ignored since finalize already ran
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      await rejection
-
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // execImmediate: ws not open after ensureConnected (line 241-242)
-  // =========================================================================
-
-  describe('execImmediate — ws disconnected between ensureConnected and send', () => {
-    it('throws "Not connected" if ws closes between connect and send', async () => {
-      const proxy = await createProxy()
-
-      // First, establish a connection
-      const warmup = proxy.exec(['version'], { priority: true })
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-      const msg0 = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({ id: msg0.id, type: 'result', payload: { output: '', exitCode: 0 } })
-      await warmup
-
-      // Now close the WS without the proxy knowing about it right away
-      // by setting readyState directly
-      activeWs!.readyState = WS_CLOSED
-      // Also null out the ws to simulate the onclose handler having fired
-      activeWs!.simulateClose()
-      await vi.advanceTimersByTimeAsync(0)
-
-      // Next exec attempt should fail due to cooldown (close sets lastConnectionFailureAt)
-      await expect(proxy.exec(['get', 'pods'], { priority: true }))
-        .rejects.toThrow('Local agent unavailable (cooldown)')
-
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // processQueue: empty queue returns early (line 211-213)
-  // =========================================================================
-
-  describe('processQueue — empty queue no-op', () => {
-    it('does nothing when queue is empty and request completes', async () => {
-      const proxy = await createProxy()
-
-      // Execute a single request — after it completes, processQueue
-      // will be called with an empty queue (covering the early return)
-      const execPromise = proxy.exec(['get', 'pods'])
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: 'done', exitCode: 0 },
-      })
-
-      const result = await execPromise
-      expect(result.output).toBe('done')
-
-      // Verify queue is empty
-      expect(proxy.getQueueStats().queued).toBe(0)
-      expect(proxy.getQueueStats().active).toBe(0)
-
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // processQueue: error propagation (line 222)
-  // =========================================================================
-
-  describe('processQueue — error propagation through queue', () => {
-    it('rejects queued request when execImmediate throws', async () => {
-      const proxy = await createProxy()
-
-      // Start a request that will fail during ensureConnected
-      const execPromise = proxy.exec(['get', 'pods'])
-      const rejection = expect(execPromise).rejects.toThrow()
-      await vi.advanceTimersByTimeAsync(0)
-
-      // Cause the connection to fail
-      activeWs!.simulateError()
-      await vi.advanceTimersByTimeAsync(0)
-
-      await rejection
-
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // getServices / getPVCs: error and parse-error branches
-  // =========================================================================
-
-  describe('getServices — error branches', () => {
-    it('throws on non-zero exitCode', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getServices('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '', exitCode: 1, error: 'forbidden' },
-      })
-
-      await expect(promise).rejects.toThrow('forbidden')
-      proxy.close()
-    })
-
-    it('throws on non-zero exitCode with no error message (fallback)', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getServices('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '', exitCode: 1 },
-      })
-
-      await expect(promise).rejects.toThrow('Failed to get services')
-      proxy.close()
-    })
-
-    it('throws on invalid JSON output', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getServices('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '{bad json', exitCode: 0 },
-      })
-
-      await expect(promise).rejects.toThrow('Failed to parse kubectl output as JSON')
-      proxy.close()
-    })
-
-    it('handles service with no ports and no clusterIP', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getServices('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: {
-          output: JSON.stringify({
-            items: [{
-              metadata: { name: 'headless', namespace: 'default' },
-              spec: { type: 'ClusterIP', clusterIP: '' },
-            }],
-          }),
-          exitCode: 0,
-        },
-      })
-
-      const svcs = await promise
-      expect(svcs[0].ports).toBe('')
-      expect(svcs[0].clusterIP).toBe('')
-      proxy.close()
-    })
-  })
-
-  describe('getPVCs — error branches', () => {
-    it('throws on non-zero exitCode', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getPVCs('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '', exitCode: 1, error: 'no access' },
-      })
-
-      await expect(promise).rejects.toThrow('no access')
-      proxy.close()
-    })
-
-    it('throws on non-zero exitCode with fallback message', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getPVCs('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '', exitCode: 1 },
-      })
-
-      await expect(promise).rejects.toThrow('Failed to get PVCs')
-      proxy.close()
-    })
-
-    it('throws on invalid JSON output', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getPVCs('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: 'not json', exitCode: 0 },
-      })
-
-      await expect(promise).rejects.toThrow('Failed to parse kubectl output as JSON')
-      proxy.close()
-    })
-
-    it('uses -A when no namespace is specified', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getPVCs('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      expect(msg.payload.args).toContain('-A')
-
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: JSON.stringify({ items: [] }), exitCode: 0 },
-      })
-      await promise
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // getPodIssues — error and parse branches
-  // =========================================================================
-
-  describe('getPodIssues — error branches', () => {
-    it('throws on non-zero exitCode', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getPodIssues('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '', exitCode: 1 },
-      })
-
-      await expect(promise).rejects.toThrow('Failed to get pods')
-      proxy.close()
-    })
-
-    it('throws on invalid JSON output', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getPodIssues('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '{nope', exitCode: 0 },
-      })
-
-      await expect(promise).rejects.toThrow('Failed to parse kubectl output as JSON')
-      proxy.close()
-    })
-
-    it('detects ErrImagePull and CreateContainerError', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getPodIssues('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: {
-          output: JSON.stringify({
-            items: [
-              {
-                metadata: { name: 'err-pull-pod', namespace: 'ns1' },
-                status: {
-                  phase: 'Pending',
-                  containerStatuses: [{
-                    restartCount: 0,
-                    state: { waiting: { reason: 'ErrImagePull' } },
-                  }],
-                },
-              },
-              {
-                metadata: { name: 'create-err-pod', namespace: 'ns1' },
-                status: {
-                  phase: 'Pending',
-                  containerStatuses: [{
-                    restartCount: 0,
-                    state: { waiting: { reason: 'CreateContainerError' } },
-                  }],
-                },
-              },
-            ],
-          }),
-          exitCode: 0,
-        },
-      })
-
-      const issues = await promise
-      expect(issues).toHaveLength(2)
-      expect(issues[0].issues).toContain('ErrImagePull')
-      expect(issues[1].issues).toContain('CreateContainerError')
-      proxy.close()
-    })
-
-    it('handles pod with Failed phase and no explicit reason', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getPodIssues('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: {
-          output: JSON.stringify({
-            items: [{
-              metadata: { name: 'generic-fail', namespace: 'default' },
-              status: { phase: 'Failed', containerStatuses: [] },
-            }],
-          }),
-          exitCode: 0,
-        },
-      })
-
-      const issues = await promise
-      expect(issues).toHaveLength(1)
-      expect(issues[0].status).toBe('Failed')
-      proxy.close()
-    })
-
-    it('handles Pending with Unschedulable but no reason string', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getPodIssues('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: {
-          output: JSON.stringify({
-            items: [{
-              metadata: { name: 'pend-pod', namespace: 'default' },
-              status: {
-                phase: 'Pending',
-                containerStatuses: [],
-                conditions: [
-                  { type: 'PodScheduled', status: 'False' }, // no reason field
-                ],
-              },
-            }],
-          }),
-          exitCode: 0,
-        },
-      })
-
-      const issues = await promise
-      expect(issues).toHaveLength(1)
-      expect(issues[0].status).toBe('Pending')
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // getEvents — error branches
-  // =========================================================================
-
-  describe('getEvents — error branches', () => {
-    it('throws on non-zero exitCode', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getEvents('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '', exitCode: 1, error: 'event error' },
-      })
-
-      await expect(promise).rejects.toThrow('event error')
-      proxy.close()
-    })
-
-    it('throws fallback message on non-zero exitCode without error', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getEvents('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '', exitCode: 1 },
-      })
-
-      await expect(promise).rejects.toThrow('Failed to get events')
-      proxy.close()
-    })
-
-    it('handles events with missing count (defaults to 1)', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getEvents('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: {
-          output: JSON.stringify({
-            items: [{
-              type: 'Normal',
-              reason: 'Pulled',
-              message: 'image pulled',
-              involvedObject: { kind: 'Pod', name: 'p1' },
-              metadata: { namespace: 'default' },
-              // no count field
-            }],
-          }),
-          exitCode: 0,
-        },
-      })
-
-      const events = await promise
-      expect(events[0].count).toBe(1)
-      proxy.close()
-    })
-
-    it('uses -A when no namespace specified', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getEvents('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      expect(msg.payload.args).toContain('-A')
-
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: JSON.stringify({ items: [] }), exitCode: 0 },
-      })
-      await promise
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // getDeployments — error and parse branches
-  // =========================================================================
-
-  describe('getDeployments — error branches', () => {
-    it('throws on non-zero exitCode', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getDeployments('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '', exitCode: 1, error: 'deploy error' },
-      })
-
-      await expect(promise).rejects.toThrow('deploy error')
-      proxy.close()
-    })
-
-    it('throws fallback message when no error field', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getDeployments('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '', exitCode: 1 },
-      })
-
-      await expect(promise).rejects.toThrow('Failed to get deployments')
-      proxy.close()
-    })
-
-    it('throws on invalid JSON output', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getDeployments('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '<<<invalid>>>', exitCode: 0 },
-      })
-
-      await expect(promise).rejects.toThrow('Failed to parse kubectl output as JSON')
-      proxy.close()
-    })
-
-    it('handles deployment with default replicas (missing spec.replicas)', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getDeployments('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: {
-          output: JSON.stringify({
-            items: [{
-              metadata: { name: 'dep1', namespace: 'default' },
-              spec: {}, // no replicas field, defaults to 1
-              status: { readyReplicas: 1 },
-            }],
-          }),
-          exitCode: 0,
-        },
-      })
-
-      const deps = await promise
-      expect(deps[0].replicas).toBe(1)
-      expect(deps[0].status).toBe('running')
-      expect(deps[0].progress).toBe(100)
-      proxy.close()
-    })
-
-    it('uses -n when namespace is specified', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getDeployments('ctx', 'my-ns')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      expect(msg.payload.args).toContain('-n')
-      expect(msg.payload.args).toContain('my-ns')
-
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: JSON.stringify({ items: [] }), exitCode: 0 },
-      })
-      await promise
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // getClusterUsage — exception catch branch (line 458-459)
-  // =========================================================================
-
-  describe('getClusterUsage — exception in exec', () => {
-    it('returns metricsAvailable=false when exec throws', async () => {
-      const proxy = await createProxy()
-      vi.spyOn(console, 'error').mockImplementation(() => {})
-
-      const usagePromise = proxy.getClusterUsage('ctx')
-      const rejection = expect(usagePromise).resolves.toEqual({
-        cpuUsageMillicores: 0,
-        memoryUsageBytes: 0,
-        metricsAvailable: false,
-      })
-      await vi.advanceTimersByTimeAsync(0)
-
-      // Cause the connection to fail
-      activeWs!.simulateError()
-      await vi.advanceTimersByTimeAsync(0)
-
-      await rejection
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // getClusterHealth — usage metrics timeout branch (line 480+)
-  // =========================================================================
-
-  describe('getClusterHealth — usage metrics timeout', () => {
-    it('continues with metricsAvailable=false when usage times out', async () => {
-      const proxy = await createProxy()
-      vi.spyOn(console, 'error').mockImplementation(() => {})
-
-      const healthPromise = proxy.getClusterHealth('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      // Respond to getNodes and getPodMetrics (sent in parallel)
-      const allMsgs = sentMessages.map(s => JSON.parse(s))
-      const nodesMsg = allMsgs.find(m => m.payload.args.includes('nodes'))!
-      const podsMsg = allMsgs.find(m => m.payload.args.includes('pods'))!
-
-      activeWs!.simulateMessage({
-        id: nodesMsg.id,
-        type: 'result',
-        payload: {
-          output: JSON.stringify({
-            items: [{
-              metadata: { name: 'n1', labels: {} },
-              status: {
-                conditions: [{ type: 'Ready', status: 'True' }],
-                allocatable: { cpu: '4', memory: '8Gi', 'ephemeral-storage': '50Gi' },
-              },
-            }],
-          }),
-          exitCode: 0,
-        },
-      })
-
-      activeWs!.simulateMessage({
-        id: podsMsg.id,
-        type: 'result',
-        payload: {
-          output: JSON.stringify({ items: [{ spec: { containers: [] } }] }),
-          exitCode: 0,
-        },
-      })
-
-      await vi.advanceTimersByTimeAsync(0)
-
-      // The "top nodes" message for getClusterUsage should be sent
-      await vi.advanceTimersByTimeAsync(100)
-
-      // DO NOT respond to the top nodes request — let the METRICS_SERVER_TIMEOUT_MS
-      // (5000ms in mock) expire so the usage metrics timeout branch fires
-      await vi.advanceTimersByTimeAsync(5000)
-
-      // Also advance past the per-request timeout for the kubectl top command
-      await vi.advanceTimersByTimeAsync(10_000)
-
-      const health = await healthPromise
-
-      // Health should still be available, just without usage metrics
-      expect(health.healthy).toBe(true)
-      expect(health.reachable).toBe(true)
-      expect(health.nodeCount).toBe(1)
-      // Usage metrics should be zero/unavailable
-      expect(health.metricsAvailable).toBe(false)
-
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // getNodes — fallback when error message is missing (line 276)
-  // =========================================================================
-
-  describe('getNodes — fallback error message', () => {
-    it('uses fallback when exitCode non-zero but no error message', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getNodes('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '', exitCode: 1 }, // no .error
-      })
-
-      await expect(promise).rejects.toThrow('Failed to get nodes')
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // getPodMetrics — parse error and fallback branches
-  // =========================================================================
-
-  describe('getPodMetrics — error branches', () => {
-    it('throws on invalid JSON', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getPodMetrics('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: 'broken', exitCode: 0 },
-      })
-
-      await expect(promise).rejects.toThrow('Failed to parse kubectl output as JSON')
-      proxy.close()
-    })
-
-    it('uses fallback error message when none provided', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getPodMetrics('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: { output: '', exitCode: 1 },
-      })
-
-      await expect(promise).rejects.toThrow('Failed to get pods')
-      proxy.close()
-    })
-
-    it('handles pods with no containers array', async () => {
-      const proxy = await createProxy()
-      const promise = proxy.getPodMetrics('ctx')
-      await vi.advanceTimersByTimeAsync(0)
-      activeWs!.simulateOpen()
-      await vi.advanceTimersByTimeAsync(0)
-
-      const msg = JSON.parse(sentMessages[0])
-      activeWs!.simulateMessage({
-        id: msg.id,
-        type: 'result',
-        payload: {
-          output: JSON.stringify({
-            items: [{ spec: {} }], // no containers
-          }),
-          exitCode: 0,
-        },
-      })
-
-      const result = await promise
-      expect(result.count).toBe(1)
-      expect(result.cpuRequestsMillicores).toBe(0)
-      expect(result.memoryRequestsBytes).toBe(0)
-      proxy.close()
-    })
-  })
-
-  // =========================================================================
-  // getBulkClusterHealth (entirely uncovered — lines 725-775)
-  // =========================================================================
-
+describe('KubectlProxy — additional coverage (edge cases)', () => {
   describe('getBulkClusterHealth', () => {
     it('processes multiple clusters and calls onProgress', async () => {
       const proxy = await createProxy()
